# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Ajloun Vision Trend Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "400":
          description: Validation error
        "409":
          description: Email already exists
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Unauthenticated
  /auth/me/accessibility:
    patch:
      operationId: updateAccessibility
      tags: [auth]
      summary: Update accessibility settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccessibilityInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
  /auth/me/profile:
    patch:
      operationId: updateProfile
      tags: [auth]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"

  # ─── INITIATIVES ───────────────────────────────────────
  /initiatives:
    get:
      operationId: listInitiatives
      tags: [initiatives]
      summary: List all initiatives
      parameters:
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: category
          in: query
          schema:
            type: ["string", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
        - name: offset
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of initiatives
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Initiative"
    post:
      operationId: createInitiative
      tags: [initiatives]
      summary: Create a new initiative
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InitiativeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Initiative"
...
  /initiatives/{id}:
    get:
      operationId: getInitiative
      tags: [initiatives]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Initiative
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Initiative"
        "404":
          description: Not found
    patch:
      operationId: updateInitiative
      tags: [initiatives]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InitiativeUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Initiative"
    delete:
      operationId: deleteInitiative
      tags: [initiatives]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /knowledge:
    get:
      operationId: listKnowledge
      tags: [knowledge]
      parameters:
        - name: category
          in: query
          schema:
            type: ["string", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KnowledgeEntry"
    post:
      operationId: createKnowledge
      tags: [knowledge]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KnowledgeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeEntry"
  /knowledge/{id}:
    get:
      operationId: getKnowledge
      tags: [knowledge]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeEntry"
    delete:
      operationId: deleteKnowledge
      tags: [knowledge]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── STATS ─────────────────────────────────────────────
...
  /admin/users/{id}:
    patch:
      operationId: adminUpdateUser
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
    delete:
      operationId: adminDeleteUser
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [email, password, fullName, fullNameAr]
      properties:
        email:
          type: string
        password:
          type: string
          minLength: 6
        fullName:
          type: string
        fullNameAr:
          type: string
        location:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthUser:
      type: object
      required: [id, email, fullName, fullNameAr, role, volunteerPoints, trainingPoints, activityPoints, totalPoints, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        fullName:
          type: string
        fullNameAr:
          type: string
        role:
          type: string
        avatarUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        isActive:
          type: boolean
        accessibilityFontLarge:
          type: boolean
        accessibilityHighContrast:
          type: boolean
        accessibilityScreenReader:
          type: boolean
        volunteerPoints:
          type: string
        trainingPoints:
          type: string
        activityPoints:
          type: string
        totalPoints:
          type: string
        createdAt:
          type: string
    AccessibilityInput:
      type: object
      properties:
        accessibilityFontLarge:
          type: boolean
        accessibilityHighContrast:
          type: boolean
        accessibilityScreenReader:
          type: boolean
    ProfileInput:
      type: object
      properties:
        fullName:
          type: string
        fullNameAr:
          type: string
        bio:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]

    # ─── INITIATIVES ──────────────────────
    Initiative:
      type: object
      required: [id, title, titleAr, description, descriptionAr, status, category, participantsCount, isFeatured, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        titleAr:
          type: string
        description:
          type: string
        descriptionAr:
          type: string
        status:
          type: string
        category:
          type: string
        imageUrl:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        participantsCount:
          type: integer
        targetParticipants:
          type: ["integer", "null"]
        progressPercent:
          type: ["integer", "null"]
        isFeatured:
          type: boolean
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        createdAt:
          type: string
    InitiativeInput:
      type: object
      required: [title, titleAr, description, descriptionAr, status, category]
      properties:
        title:
          type: string
        titleAr:
          type: string
        description:
          type: string
        descriptionAr:
          type: string
        status:
          type: string
        category:
          type: string
        imageUrl:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        targetParticipants:
          type: ["integer", "null"]
        progressPercent:
          type: ["integer", "null"]
        isFeatured:
          type: boolean
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
    InitiativeUpdate:
      type: object
      properties:
        title:
          type: string
        titleAr:
          type: string
        description:
          type: string
        descriptionAr:
          type: string
        status:
          type: string
        category:
          type: string
        imageUrl:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        targetParticipants:
          type: ["integer", "null"]
        progressPercent:
          type: ["integer", "null"]
        isFeatured:
          type: boolean
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
...
    KnowledgeEntry:
      type: object
      required: [id, title, titleAr, content, contentAr, category, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        titleAr:
          type: string
        content:
          type: string
        contentAr:
          type: string
        category:
          type: string
        order:
          type: integer
        createdAt:
          type: string
    KnowledgeInput:
      type: object
      required: [title, titleAr, content, contentAr, category]
      properties:
        title:
          type: string
        titleAr:
          type: string
        content:
          type: string
        contentAr:
          type: string
        category:
          type: string
        order:
          type: integer

    # ─── STATS ────────────────────────────
...
    AdminUserUpdate:
      type: object
      properties:
        role:
          type: string
        isActive:
          type: boolean
        fullName:
          type: string
        fullNameAr:
          type: string
        volunteerPoints:
          type: string
        trainingPoints:
          type: string
        activityPoints:
          type: string
        totalPoints:
          type: string